server/requirements.txt
certifi==2022.5.18.1
click==8.1.3
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
opencv-python==4.5.5.64
pandas==1.4.2
Pillow==9.1.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
requests==2.27.1